event-broker: Bump libsystemd dependency to 0.7.2
Since libsystemd 0.7.1, a mismatch of PIDs which owns the socket and
passing the FD is no longer treated as an error. As this is the case
for a standalone server, this removes the error handling and
adds the fallback logic under the condition where receive_descriptors
returns an empty array.

Signed-off-by: Daiki Ueno <dueno@redhat.com>